feat: Implement Entity, Repository, and Service layers with database integration
- Add Debt and DebtInstallment entities with JPA annotations
- Add DebtStatus enum (ACTIVE, PAID_OFF)
- Implement DebtRepository and DebtInstallmentRepository with custom queries
- Implement DebtService for debt CRUD operations
- Implement InstallmentService for installment management
- Implement DashboardService for statistics and aggregations
- Update AdminController and DashboardController to use services instead of sample data
- Fix database schema: change amount columns from INTEGER to BIGINT
- Add null safety checks and eager loading for relationships
- All operations now persist to PostgreSQL database

+++ src/main/resources/db/migration/V1__init.sql
@@ -2,7 +2,7 @@
 CREATE TABLE debts (
     id BIGSERIAL PRIMARY KEY,
     title VARCHAR(200) NOT NULL,
-    total_amount INTEGER NOT NULL CHECK (total_amount > 0),
+    total_amount BIGINT NOT NULL CHECK (total_amount > 0),
     installment_count INTEGER NOT NULL CHECK (installment_count > 0),
     start_date DATE NOT NULL,
     interest_rate DECIMAL(5, 2) NULL CHECK (interest_rate >= 0),
@@ -16,7 +16,7 @@ CREATE TABLE debt_installments (
     id BIGSERIAL PRIMARY KEY,
     debt_id BIGINT NOT NULL REFERENCES debts(id) ON DELETE CASCADE,
     installment_number INTEGER NOT NULL CHECK (installment_number > 0),
-    amount INTEGER NOT NULL CHECK (amount > 0),
+    amount BIGINT NOT NULL CHECK (amount > 0),
     due_date DATE NOT NULL,
     paid BOOLEAN NOT NULL DEFAULT FALSE,
     paid_at DATE NULL,
